_algorithms_challenges/pybites/bitesofpy-master/!201-300/228/gravatar.py
import hashlib

GRAVATAR_URL = ("https://www.gravatar.com/avatar/"
                "{hashed_email}?s={size}&r=g&d=robohash")


def create_gravatar_url(email: str, size=200):
    """Use GRAVATAR_URL above to create a gravatar URL.

       You need to create a hash of the email passed in.

       PHP example: https://en.gravatar.com/site/implement/hash/

       For Python check hashlib check out (md5 / hexdigest):
       https://docs.python.org/3/library/hashlib.html#hashlib.hash.hexdigest
    """
    return GRAVATAR_URL.format(hashed_email=hashlib.md5(email.strip().lower().encode('utf-8')).hexdigest(),
                               size=size)
